+++ packages/core/src/odata/common/edm-types.ts
@@ -15,7 +15,8 @@ export type EdmTypeSameConverters =
   | 'Edm.SByte'
   | 'Edm.Binary'
   | 'Edm.Guid'
-  | 'Edm.Byte';
+  | 'Edm.Byte'
+  | 'Edm.Any'; // Represents currently unspupported edm types like Edm.Geography.
 
 export type EdmTypeDifferentConverters = 'Edm.DateTimeOffset';