+++ packages/core/test/test-util/test-services/v4/test-service/TestComplexType.ts
@@ -6,9 +6,9 @@
 import { Moment } from 'moment';
 import { BigNumber } from 'bignumber.js';
 import { TestNestedComplexType, TestNestedComplexTypeField } from './TestNestedComplexType';
-import { CollectionField, ComplexTypeBigNumberPropertyField, ComplexTypeBooleanPropertyField, ComplexTypeDatePropertyField, ComplexTypeField, ComplexTypeNumberPropertyField, ComplexTypeStringPropertyField, ComplexTypeTimePropertyField, ConstructorOrField, Entity, FieldType, PropertyMetadata, Time, deserializeComplexType } from '../../../../../src/v4';
 import { ComplexTypeEnumPropertyField } from '../../../../../src/odata/common/selectable/enum-field';
 import { TestEnumType } from './TestEnumType';
+import { CollectionField, ComplexTypeAnyPropertyField, ComplexTypeBigNumberPropertyField, ComplexTypeBooleanPropertyField, ComplexTypeDatePropertyField, ComplexTypeField, ComplexTypeNumberPropertyField, ComplexTypeStringPropertyField, ComplexTypeTimePropertyField, ConstructorOrField, Entity, FieldType, PropertyMetadata, Time, deserializeComplexType } from '../../../../../src/v4';
 
 /**
  * TestComplexType
@@ -90,6 +90,11 @@ export interface TestComplexType {
   sByteProperty?: number;
 
   enumProperty?: TestEnumType;
+  /**
+   * Geography Point Property.
+   * @nullable
+   */
+  geographyPointProperty?: any;
   /**
    * Complex Type Property.
    * @nullable
@@ -201,6 +206,11 @@ export class TestComplexTypeField<EntityT extends Entity> extends ComplexTypeFie
   sByteProperty: ComplexTypeNumberPropertyField<EntityT> = new ComplexTypeNumberPropertyField('SByteProperty', this, 'Edm.SByte');
 
   enumProperty: ComplexTypeEnumPropertyField<EntityT> = new ComplexTypeEnumPropertyField('EnumProperty', this, 'Edm.Enum');
+  /**
+   * Representation of the [[TestComplexType.geographyPointProperty]] property for query construction.
+   * Use to reference this property in query operations such as 'filter' in the fluent request API.
+   */
+  geographyPointProperty: ComplexTypeAnyPropertyField<EntityT> = new ComplexTypeAnyPropertyField('GeographyPointProperty', this, 'Edm.Any');
   /**
    * Representation of the [[TestComplexType.complexTypeProperty]] property for query construction.
    * Use to reference this property in query operations such as 'filter' in the fluent request API.
@@ -316,6 +326,11 @@ export namespace TestComplexType {
     originalName: 'SByteProperty',
     name: 'sByteProperty',
     type: 'Edm.SByte',
+    isCollection: false
+  }, {
+    originalName: 'GeographyPointProperty',
+    name: 'geographyPointProperty',
+    type: 'Edm.Any',
     isCollection: false
   }, {
     originalName: 'ComplexTypeProperty',