+++ packages/generator/src/generator-utils.ts
@@ -42,7 +42,7 @@ export function isNullableParameter(parameter: any) {
   return !!parameter['Nullable'] && parameter['Nullable'] !== 'false';
 }
 
-type EdmTypeMapping = { [key in EdmTypeShared<'any'>]: string | undefined };
+export type EdmTypeMapping = { [key in EdmTypeShared<'any'>]: string | undefined };
 
 const edmToTsTypeMapping: EdmTypeMapping = {
   'Edm.String': 'string',
@@ -59,6 +59,7 @@ const edmToTsTypeMapping: EdmTypeMapping = {
   'Edm.SByte': 'number',
   'Edm.DateTimeOffset': 'Moment',
   'Edm.Binary': 'string',
+  'Edm.Any': 'any',
 
   // OData v2 specific
   'Edm.DateTime': 'Moment',
@@ -86,6 +87,7 @@ const edmToFieldTypeMapping: EdmTypeMapping = {
   'Edm.SByte': 'NumberField',
   'Edm.DateTimeOffset': 'DateField',
   'Edm.Binary': 'BinaryField',
+  'Edm.Any': 'AnyField',
 
   // OData v2 specific
   'Edm.DateTime': 'DateField',
@@ -105,21 +107,45 @@ const fieldTypeToComplexPropertyTypeMapping = {
   DateField: 'ComplexTypeDatePropertyField',
   NumberField: 'ComplexTypeNumberPropertyField',
   StringField: 'ComplexTypeStringPropertyField',
-  TimeField: 'ComplexTypeTimePropertyField'
+  TimeField: 'ComplexTypeTimePropertyField',
+  AnyField: 'ComplexTypeAnyPropertyField'
 };
 
-export function edmToTsType(edmType: string): string | undefined {
-  if (edmToTsTypeMapping[edmType]) {
-    return edmToTsTypeMapping[edmType];
+export function getFallbackEdmTypeIfNeeded(
+  edmType: string
+): EdmTypeShared<any> {
+  if (edmType in edmToTsTypeMapping) {
+    return edmType as EdmTypeShared<any>;
   }
+  logger.warn(
+    `The type ${edmType} is currently not supported by the sdk. Type "any" is used as fallback.`
+  );
+  return 'Edm.Any';
+}
+
+export function edmToTsType(edmType: string): string {
+  const tsType = edmToTsTypeMapping[edmType];
+  if (!tsType) {
+    throw new Error(`No ts type found for edm type: ${edmType}`);
+  }
+  return tsType;
 }
 
 export function edmToFieldType(edmType: string): string {
-  return edmToFieldTypeMapping[edmType];
+  const fieldType = edmToFieldTypeMapping[edmType];
+  if (!fieldType) {
+    throw new Error(`No field type found for edm type: ${edmType}`);
+  }
+  return fieldType;
 }
 
 export function edmToComplexPropertyType(edmType: string): string {
-  return fieldTypeToComplexPropertyTypeMapping[edmToFieldType(edmType)];
+  const fieldType =
+    fieldTypeToComplexPropertyTypeMapping[edmToFieldType(edmType)];
+  if (!fieldType) {
+    throw new Error(`No complex field type found for edm type: ${edmType}`);
+  }
+  return fieldType;
 }
 
 export function forceArray(obj: any): any[] {